+++ CHANGELOG.md
@@ -9,6 +9,8 @@ Format: `<description> (<contributor>, <commit>)`
 ## Added
 
 - Find notes with missing backlinks using `zk list --missing-backlink`
+- LSP diagnostic for missing backlinks when other notes link to current note
+  without reciprocal links (by @WhyNotHugo)
 
 ## Fixed